+++ packages/zent/src/datetimepicker/quarter/QuarterPanelBody.js
@@ -1,31 +1,31 @@
 import React, { Component, PureComponent } from 'react';
 import classNames from 'classnames';
 import PanelCell from '../common/PanelCell';
-import { getQuaterFromDate } from '../utils/date';
+import { getQuarterFromDate } from '../utils/date';
 
 const ROW = 2;
 const COL = 2;
 
-export default class QuaterPanelBody extends (PureComponent || Component) {
-  getQuaters() {
+export default class QuarterPanelBody extends (PureComponent || Component) {
+  getQuarters() {
     const { disabledDate, selected, i18n } = this.props;
-    const quaters = [];
+    const quarters = [];
     let index = 0;
     for (let rowIndex = 0; rowIndex < ROW; rowIndex++) {
-      quaters[rowIndex] = [];
+      quarters[rowIndex] = [];
       for (let colIndex = 0; colIndex < COL; colIndex++) {
-        const isSelected = selected && getQuaterFromDate(selected) === index;
+        const isSelected = selected && getQuarterFromDate(selected) === index;
         const isDisabled = disabledDate && disabledDate(index);
         const className = classNames({
-          'panel__cell quater-panel__cell': true,
+          'panel__cell quarter-panel__cell': true,
           'panel__cell--current': false,
           'panel__cell--selected': isSelected,
           'panel__cell--disabled': isDisabled
         });
-        quaters[rowIndex][colIndex] = {
-          text: i18n.panel.quaterNames[index],
+        quarters[rowIndex][colIndex] = {
+          text: i18n.panel.quarterNames[index],
           value: index,
-          title: i18n.panel.quaterNames[index],
+          title: i18n.panel.quarterNames[index],
           className,
           isDisabled
         };
@@ -33,16 +33,16 @@ export default class QuaterPanelBody extends (PureComponent || Component) {
       }
     }
 
-    return quaters;
+    return quarters;
   }
 
   render() {
     const { onSelect } = this.props;
-    const quaters = this.getQuaters();
+    const quarters = this.getQuarters();
 
     return (
-      <div className="quater-table panel-table">
-        <PanelCell onSelect={onSelect} cells={quaters} />
+      <div className="quarter-table panel-table">
+        <PanelCell onSelect={onSelect} cells={quarters} />
       </div>
     );
   }